+++ tests/testthat/test_cpi.R
@@ -1,11 +1,19 @@
 context("Inflators return correct results")
 
+test_that("Default from_fy and to_fy", {
+  expect_warning(cpi_inflator(), 
+                 regexp = "`from_fy` and `to_fy` are missing, using previous and current financial years respectively")
+  expect_equal(suppressWarnings(cpi_inflator()),
+               cpi_inflator(from_fy = yr2fy(year(Sys.Date()) - 1),
+                              to_fy = date2fy(Sys.Date())))
+})
+
 test_that("Errors", {
-  expect_error(cpi_inflator(from_fy = NA, to_fy = "2013-14"), 
-               regexp = "`from_fy` contained NAs. Remove NAs before applying.", 
+  expect_error(cpi_inflator(to_fy = "2013-14"), 
+               regexp = "`from_fy` is missing", 
                fixed = TRUE)
-  expect_error(cpi_inflator(to_fy = NA, from_fy = "2013-14"), 
-               regexp = "`to_fy` contained NAs. Remove NAs before applying.",
+  expect_error(cpi_inflator(from_fy = "2013-14"), 
+               regexp = "`to_fy` is missing",
                fixed = TRUE)
   expect_error(cpi_inflator(from_fy = "2013-14", to_fy = "2040-41",
                             allow.projection = FALSE),